Refactored each system owns memory..
Gave each system the responsibility to allocate, manage and pop its own memory arenas.
This allowed for de-coupling the renderer from the rest of the system.
Also, rewrote renderer so that bitmap is also update(popped and re-allocated) every time the user changes the height or width of the window.